src/auth/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.3
